+ room dep;separate TFLiteModelData
