+++ client/fedcampus/build.gradle
@@ -71,6 +71,13 @@ dependencies {
     implementation "io.grpc:grpc-stub:$grpc_version"
     implementation 'javax.annotation:javax.annotation-api:1.3.2'
 
+    // Room.
+    def room_version = "2.5.1"
+    implementation("androidx.room:room-runtime:$room_version")
+    annotationProcessor("androidx.room:room-compiler:$room_version")
+    implementation("androidx.room:room-ktx:$room_version")
+    testImplementation("androidx.room:room-testing:$room_version")
+
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
